+++ process_image.h
@@ -1,7 +1,7 @@
 #ifndef PROCESS_IMAGE_H
 #define PROCESS_IMAGE_H
 
-float get_offset_from_center(void);
+int16_t get_offset_from_center(void);
 void process_image_start(void);
 
 #endif /* PROCESS_IMAGE_H */
